some code to try to get coy of default report working
I didn't get the post processing to run on this to produce a summary for the scenario. I want to get a very simple example working. I've thought of just writing a script not using the SDK that grabs what it wants out of the out.osw files, but I'd also be happy to make JSON and CSV mathing approach of default report.